Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/html/dom/dynamic-markup-insertion/opening-the-input-stream/015-expected.txt


FAIL global scope unchanged assert_equals: expected 1 but got 2
PASS window object changed 
PASS this is the window object 
